+++ apps/frontend/src/modules/categories/categories.ts
@@ -10,5 +10,6 @@ const categoriesApi = new CategoriesApi({
 	storage,
 });
 
+export { type CategoriesGetRequestQueryDto } from "./libs/types/types.js";
 export { actions, reducer } from "./slices/categories.js";
 export { categoriesApi };