+++ apps/frontend/src/modules/quiz/slices/actions.ts
@@ -2,6 +2,7 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 
 import { type AsyncThunkConfig } from "~/libs/types/types.js";
 import {
+	type CategoriesGetRequestQueryDto,
 	type QuizAnswersRequestDto,
 	type QuizQuestionDto,
 	type QuizScoresGetAllResponseDto,
@@ -32,6 +33,16 @@ const getAllQuestions = createAsyncThunk<
 	return await quizApi.getAllQuestions();
 });
 
+const getQuestionsByCategoryIds = createAsyncThunk<
+	{ items: QuizQuestionDto[][] },
+	CategoriesGetRequestQueryDto,
+	AsyncThunkConfig
+>(`${sliceName}/get-questions-by-category-ids`, async (payload, { extra }) => {
+	const { quizApi } = extra;
+
+	return await quizApi.getQuestionsByCategoryIds(payload.categoryIds);
+});
+
 const saveAnswers = createAsyncThunk<
 	QuizUserAnswerDto[],
 	QuizAnswersRequestDto,
@@ -52,4 +63,10 @@ const getScores = createAsyncThunk<
 	return await quizApi.getScores();
 });
 
-export { editScores, getAllQuestions, getScores, saveAnswers };
+export {
+	editScores,
+	getAllQuestions,
+	getQuestionsByCategoryIds,
+	getScores,
+	saveAnswers,
+};