+++ apps/frontend/src/modules/tasks/libs/constants/constants.ts
@@ -1,5 +1,3 @@
 const NOT_FOUND_INDEX = -1;
 
-const SINGLE_ELEMENT = 1;
-
-export { NOT_FOUND_INDEX, SINGLE_ELEMENT };
+export { NOT_FOUND_INDEX };