+++ apps/frontend/src/modules/tasks/libs/types/task-update-payload.type.ts
@@ -2,7 +2,6 @@ import { type TaskUpdateRequestDto } from "./types.js";
 
 type TaskUpdatePayload = {
 	id: number;
-	task: Partial<TaskUpdateRequestDto>;
-};
+} & Partial<TaskUpdateRequestDto>;
 
 export { type TaskUpdatePayload };