+++ apps/frontend/src/pages/root/components/user-wheel/user-wheel.tsx
@@ -2,7 +2,7 @@ import {
 	BalanceWheelChart,
 	Button,
 	Loader,
-	ScoresEditModal,
+	Switch,
 } from "~/libs/components/components.js";
 import {
 	useAppDispatch,
@@ -13,15 +13,28 @@ import {
 } from "~/libs/hooks/hooks.js";
 import { actions as quizActions } from "~/modules/quiz/quiz.js";
 
+import { CircularProgress } from "../circular-progress/circular-progress.js";
+import {
+	RetakeQuizModal,
+	ScoresEditModal,
+} from "./libs/components/components.js";
+import { type WheelEditMode } from "./libs/types/types.js";
 import styles from "./styles.module.css";
 
+const NO_SCORES_COUNT = 0;
+
 const UserWheel: React.FC = () => {
-	const NO_SCORES_COUNT = 0;
-
 	const dispatch = useAppDispatch();
-	const { dataStatus, scores } = useAppSelector((state) => state.quiz);
+	const { completionTasksPercentage, dataStatus, scores } = useAppSelector(
+		({ auth, quiz }) => ({
+			completionTasksPercentage: auth.user?.completionTasksPercentage,
+			dataStatus: quiz.dataStatus,
+			scores: quiz.scores,
+		}),
+	);
 	const [isEditingModalOpen, setIsEditingModalOpen] = useState<boolean>(false);
-
+	const [percentage, setPercentage] = useState<number>(NO_SCORES_COUNT);
+	const [editMode, setEditMode] = useState<WheelEditMode>("manual");
 	const isLoading = dataStatus === "pending";
 
 	const chartData = scores.map((score) => {
@@ -43,23 +56,65 @@ const UserWheel: React.FC = () => {
 		setIsEditingModalOpen(false);
 	}, []);
 
+	const handleModeToggle = useCallback(() => {
+		setEditMode((previousState) => {
+			return previousState === "manual" ? "retake_quiz" : "manual";
+		});
+	}, []);
+
 	useEffect(() => {
 		void dispatch(quizActions.getScores());
 	}, [dispatch]);
+
+	useEffect(() => {
+		if (completionTasksPercentage) {
+			setPercentage(completionTasksPercentage);
+		}
+	}, [completionTasksPercentage]);
+
+	const handleGetModal = (mode: WheelEditMode): React.ReactNode => {
+		switch (mode) {
+			case "manual": {
+				return (
+					<ScoresEditModal data={scores} onSaveChanges={handleFinishEditing} />
+				);
+			}
+
+			case "retake_quiz": {
+				return <RetakeQuizModal />;
+			}
+
+			default: {
+				return null;
+			}
+		}
+	};
 
 	return (
 		<div className={styles["container"]}>
 			<div className={styles["header"]}>
 				<h4 className={styles["header-text"]}>{headerText}</h4>
+				{isEditingModalOpen && (
+					<div className={styles["switch-container"]}>
+						<Switch
+							currentMode={editMode}
+							leftButtonProperties={{ label: "Edit manually", mode: "manual" }}
+							onToggleMode={handleModeToggle}
+							rightButtonProperties={{
+								label: "Retake quiz",
+								mode: "retake_quiz",
+							}}
+						/>
+					</div>
+				)}
 			</div>
 			<div className={styles["content-wrapper"]}>
 				{scores.length > NO_SCORES_COUNT && (
 					<BalanceWheelChart data={chartData} />
 				)}
-				{isEditingModalOpen && (
-					<ScoresEditModal data={scores} onSaveChanges={handleFinishEditing} />
-				)}
+				{isEditingModalOpen && handleGetModal(editMode)}
 			</div>
+			<CircularProgress percentage={percentage} />
 			{!isEditingModalOpen && (
 				<div className={styles["button-wrapper"]}>
 					<Button