+++ apps/frontend/src/pages/tasks/libs/components/task-card/task-card.tsx
@@ -1,31 +1,91 @@
+import { Button } from "~/libs/components/components.js";
 import { getValidClassNames } from "~/libs/helpers/helpers.js";
+import { useCallback } from "~/libs/hooks/hooks.js";
 import { type TaskDto } from "~/modules/tasks/tasks.js";
 
-import { Category, Deadline } from "../components.js";
+import { TaskStatus } from "../../enums/enums.js";
+import { Category, Deadline, PastTaskStatus } from "../components.js";
 import styles from "./styles.module.css";
 
 type Properties = {
+	onComplete?: (id: number) => void;
 	onExpire?: (expiredTask: TaskDto) => void;
+	onSkip?: (id: number) => void;
 	task: TaskDto;
 	variant?: "active" | "expired";
 };
 
 const TaskCard: React.FC<Properties> = ({
+	onComplete = (): void => {},
 	onExpire,
+	onSkip = (): void => {},
 	task,
 	variant = "active",
 }: Properties) => {
+	const handleSkip = useCallback(() => {
+		onSkip(task.id);
+	}, [task, onSkip]);
+
+	const handleComplete = useCallback(() => {
+		onComplete(task.id);
+	}, [task, onComplete]);
+
+	const areActionsDisabled = variant === "expired";
+	const buttonsContainerClass = getValidClassNames(
+		styles["buttons-container"],
+		areActionsDisabled && styles["buttons-container-expired"],
+	);
+	const isActive = task.status === TaskStatus.CURRENT;
+
 	return (
 		<div
 			className={getValidClassNames(styles["card"], styles[`card-${variant}`])}
 		>
 			<div className={styles["card-header"]}>
 				<Category categoryName={task.category} variant={variant} />
-				<Deadline onExpire={onExpire} task={task} />
+				{isActive && <Deadline onExpire={onExpire} task={task} />}
 			</div>
 			<div className={styles["card-body"]}>
 				<h4 className={styles["title"]}>{task.label}</h4>
 				<div className={styles["text"]}>{task.description}</div>
+			</div>
+			<div className={styles["card-footer"]}>
+				<div
+					className={getValidClassNames(
+						styles["divider"],
+						styles[`divider-${variant}`],
+					)}
+				/>
+				<div className={buttonsContainerClass}>
+					{isActive ? (
+						<>
+							<div className={styles["button-container"]}>
+								<Button
+									iconName="closeSmall"
+									isDisabled={areActionsDisabled}
+									label="Skip the task"
+									onClick={handleSkip}
+									type="button"
+									variant="action"
+								/>
+							</div>
+							<div className={styles["button-container"]}>
+								<Button
+									iconName="checkBlack"
+									isDisabled={areActionsDisabled}
+									label="Mark complete"
+									onClick={handleComplete}
+									type="button"
+									variant="action"
+								/>
+							</div>
+						</>
+					) : (
+						<div>
+							<PastTaskStatus status={task.status} />
+						</div>
+					)}
+				</div>
 			</div>
 		</div>
 	);