+++ apps/frontend/src/pages/tasks/tasks.tsx
@@ -1,4 +1,4 @@
-import { Loader } from "~/libs/components/components.js";
+import { Loader, Switch } from "~/libs/components/components.js";
 import { DataStatus } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
@@ -7,10 +7,12 @@ import {
 	useEffect,
 	useState,
 } from "~/libs/hooks/hooks.js";
+import { type ValueOf } from "~/libs/types/types.js";
 import { actions as taskActions, type TaskDto } from "~/modules/tasks/tasks.js";
 
 import { ExpiredTasksModal, TaskCard } from "./libs/components/components.js";
 import { NO_EXPIRED_TASKS, ONE_MINUTE } from "./libs/constants/constants.js";
+import { TasksMode, TaskStatus } from "./libs/enums/enums.js";
 import { getTimeLeft } from "./libs/helpers/helpers.js";
 import styles from "./styles.module.css";
 
@@ -25,9 +27,9 @@ const Tasks: React.FC = () => {
 	const [expiredTasks, setExpiredTasks] = useState<TaskDto[]>([]);
 	const [activeTasks, setActiveTasks] = useState<TaskDto[]>([]);
 
-	useEffect(() => {
-		void dispatch(taskActions.getCurrentTasks());
-	}, [dispatch]);
+	const [mode, setMode] = useState<ValueOf<typeof TasksMode>>(
+		TasksMode.CURRENT,
+	);
 
 	const handleTaskExpiration = useCallback(
 		(expiredTask: TaskDto) => {
@@ -60,6 +62,42 @@ const Tasks: React.FC = () => {
 		setActiveTasks(active);
 	}, [tasks]);
 
+	useEffect(() => {
+		void dispatch(taskActions.getCurrentTasks());
+	}, [dispatch]);
+
+	useEffect(() => {
+		if (mode === TasksMode.CURRENT) {
+			void dispatch(taskActions.getCurrentTasks());
+		}
+
+		if (mode === TasksMode.PAST) {
+			void dispatch(taskActions.getPastTasks());
+		}
+	}, [dispatch, mode]);
+
+	const handleModeToggle = useCallback(() => {
+		setMode((previousState) => {
+			return previousState === TasksMode.CURRENT
+				? TasksMode.PAST
+				: TasksMode.CURRENT;
+		});
+	}, []);
+
+	const handleSkip = useCallback(
+		(id: number): void => {
+			void dispatch(taskActions.update({ id, status: TaskStatus.SKIPPED }));
+		},
+		[dispatch],
+	);
+
+	const handleComplete = useCallback(
+		(id: number): void => {
+			void dispatch(taskActions.update({ id, status: TaskStatus.COMPLETED }));
+		},
+		[dispatch],
+	);
+
 	const isLoading = dataStatus === DataStatus.PENDING;
 
 	return (
@@ -69,17 +107,34 @@ const Tasks: React.FC = () => {
 			)}
 			<h4 className={styles["title"]}>My Tasks</h4>
 			<div className={styles["board"]}>
-				{isLoading ? (
-					<Loader />
-				) : (
-					activeTasks.map((task) => (
-						<TaskCard
-							key={task.id}
-							onExpire={handleTaskExpiration}
-							task={task}
+				<div className={styles["board-header"]}>
+					<div className={styles["switch-container"]}>
+						<Switch
+							currentMode={mode}
+							leftButtonProperties={{
+								label: "Active",
+								mode: TasksMode.CURRENT,
+							}}
+							onToggleMode={handleModeToggle}
+							rightButtonProperties={{ label: "Past", mode: TasksMode.PAST }}
 						/>
-					))
-				)}
+					</div>
+				</div>
+				<div className={styles["cards-container"]}>
+					{isLoading ? (
+						<Loader />
+					) : (
+						activeTasks.map((task) => (
+							<TaskCard
+								key={task.id}
+								onComplete={handleComplete}
+								onExpire={handleTaskExpiration}
+								onSkip={handleSkip}
+								task={task}
+							/>
+						))
+					)}
+				</div>
 			</div>
 		</>
 	);